+++ examples/skill-sample-nodejs-deviceaddressapi/helloworld.spec.ts
@@ -6,7 +6,7 @@ import { AlexaTest, LaunchRequestBuilder, SkillSettings } from '../../lib';
 import { handler as skillHandler } from './helloworld';
 
 // initialize the testing framework
-const skillSettings : SkillSettings = {
+const skillSettings: SkillSettings = {
     appId: 'amzn1.ask.skill.00000000-0000-0000-0000-000000000000',
     userId: 'amzn1.ask.account.VOID',
     deviceId: 'amzn1.ask.device.VOID',
@@ -18,21 +18,26 @@ const alexaTest = new AlexaTest(skillHandler, skillSettings);
 describe('Hello World Skill Profile API', () => {
 
     describe('LaunchRequest', () => {
+        const cardText = `Short Address: countryCode: US, postalCode: 98109\n` +
+            `Full Address: line1: 410 Terry Ave North, line2: Apt 3, line3: , ` +
+            `city: Seattle, countryCode: US, districtOrCounty: , ` +
+            `postalCode: 98109, stateOrRegion: WA`
         alexaTest.test([
             {
                 request: new LaunchRequestBuilder(skillSettings).build(),
-                withProfile: {
-                    givenName: 'John',
-                    name: 'Smith',
-                    email: 'john@smith.com',
-                    mobileNumber: '+1234567890',
-                    distanceUnits: 'METRIC',
-                    temperatureUnits: 'CELSIUS',
-                    timeZone: 'Africa/Abidjan'
+                withDeviceAddress: {
+                    stateOrRegion: "WA",
+                    city: "Seattle",
+                    countryCode: "US",
+                    postalCode: "98109",
+                    addressLine1: "410 Terry Ave North",
+                    addressLine2: "Apt 3",
+                    addressLine3: "",
+                    districtOrCounty: ""
                 },
-                says: 'Hello, John Smith. Your e-mail is john@smith.com and your phone number is +1234567890',
+                says: 'Hello, world!',
                 hasCardTitle: 'Hello World',
-                hasCardContent: 'Distance Unit: METRIC\nTemperature Unit: CELSIUS\nTime Zone: Africa/Abidjan',
+                hasCardContent: cardText,
                 repromptsNothing: true,
                 shouldEndSession: true,
             },
@@ -43,8 +48,8 @@ describe('Hello World Skill Profile API', () => {
         alexaTest.test([
             {
                 request: new LaunchRequestBuilder(skillSettings).build(),
-                says: 'Hello, world! I am not allowed to view your profile.',
-                hasAskForPermissionsConsentCard: ['alexa::profile:name:read', 'alexa::profile:email:read', 'alexa::profile:mobile_number:read'],
+                says: 'Hello, world! I am not allowed to view your address.',
+                hasAskForPermissionsConsentCard: ['alexa:devices:all:address:country_and_postal_code:read', 'alexa::devices:all:address:full:read'],
                 repromptsNothing: true,
                 shouldEndSession: true,
             },